trigger heart even there is no http requests (use threading.Timer

+++ tools/common/file_storage.py
@@ -77,11 +77,14 @@ class TFileStorage:
         if self.header_archive_timeout is not None:
             assert self.archive_folder is not None
 
-    def sync_header_with_archive(self):
+    def sync_with_archive(self, force=False):
         if self.archive_folder is not None:
-            self.logger.debug("copy {} to {}".format(self.dbm_path, self.archive_folder))
-            self.header_last_sync_timestamp = time.time()
-            shutil.copy(self.dbm_path, self.archive_folder)
+            if self.header_archive_timeout is not None:
+                if force  or (time.time() - self.header_last_sync_timestamp > self.header_archive_timeout):
+                    self.logger.debug("copy {} to {}".format(self.dbm_path, self.archive_folder))
+                    self.header_last_sync_timestamp = time.time()
+                    shutil.copy(self.dbm_path, self.archive_folder)
+                    shutil.copy(self.bin_files[-1].name, self.archive_folder)
 
     def write_key_to_dbm(self, key, value):
         self.saved_file_params[key] = value
@@ -93,9 +96,7 @@ class TFileStorage:
             self.saved_file_params.sync()
             self.write_without_sync_count = 0
 
-        if self.header_archive_timeout is not None:
-            if time.time() - self.header_last_sync_timestamp > self.header_archive_timeout:
-                self.sync_header_with_archive()
+        self.sync_with_archive()
 
     def get_all_keys(self):
         k = self.saved_file_params.firstkey()
@@ -172,17 +173,21 @@ class TFileStorage:
     def create_new_bin_file(self):
         self.bin_files[-1].close()
         last_file_path = self.get_bin_file_path(len(self.bin_files) - 1)
-        if self.archive_folder is not  None:
-            basename = os.path.basename(last_file_path)
-            shutil.move(last_file_path, self.archive_folder)
-            os.symlink(os.path.join(self.archive_folder, basename), last_file_path)
+        if self.archive_folder is not None:
+            output_file = os.path.join(self.archive_folder, os.path.basename(last_file_path))
+            if os.path.exists(output_file):
+                if os.path.getsize(last_file_path) < os.path.getsize(output_file):
+                    raise Exception("cannot move {} to {}, since the last folder comtains a bigger file".format(last_file_path, self.archive_folder))
+                os.unlink(output_file)
+            shutil.move(last_file_path, output_file)
+            os.symlink(output_file, last_file_path)
             if not os.path.islink(last_file_path):
                 raise Exception("{} must be a symbolic link".format(last_file_path))
-            self.sync_header_with_archive()
 
         self.bin_files[-1] = open(last_file_path, "rb")
 
         self.bin_files.append(open(self.get_bin_file_path(len(self.bin_files)), "ab+"))
+        self.sync_with_archive(force=True)
 
     def write_repeat_header_to_bin_file(self, file_bytes, file_extension, output_bin_file):
         # these headers are needed if the main dbm is lost
